% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{augmentMultiple}
\alias{augmentMultiple}
\title{Create the data for the plots}
\usage{
augmentMultiple(...)
}
\arguments{
\item{...}{A list of models}
}
\description{
Augment the data from a model output to be in a form suitable for ggplot
}
