% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ggtimedistCDF}
\alias{ggtimedistCDF}
\title{Plot the timedist CDF}
\usage{
ggtimedistCDF(...)
}
\arguments{
\item{...}{A model (or a list of models) of class \code{timedist}.}
}
\description{
Given a model (or models) of class \code{timedist}, produce a cumulative
distribution plot for each of them.
}

