% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ggtimedistPDF}
\alias{ggtimedistPDF}
\title{Plot the timedist PDF}
\usage{
ggtimedistPDF(..., xlim = NULL)
}
\arguments{
\item{...}{A model (or a list of models) of class \code{timedist}.}

\item{xlim}{The x limits (x1, x2) of the plot.}
}
\description{
Given a model (or models) of class \code{timedist}, produce a probability density function plot
for each of them.
}

