% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance.timedist.R
\name{glance.timedist}
\alias{glance.timedist}
\title{Construct a single row summary "glance" of a timedist model}
\usage{
\method{glance}{timedist}(x, ...)
}
\arguments{
\item{x}{An object of class \code{timedist}.}

\item{...}{Additional arguments (not used).}
}
\value{
\code{glance} returns one row with the columns
  \item{sigma}{the square root of the estimated residual variance}
  \item{isConv}{whether the fit successfully converged}
  \item{finTol}{the achieved convergence tolerance}
  \item{logLik}{the data's log-likelihood under the model}
  \item{AIC}{the Akaike Information Criterion}
  \item{BIC}{the Bayesian Information Criterion}
  \item{deviance}{deviance}
  \item{df.residual}{residual degrees of freedom}
  \item{RSS}{corrected residual sum of squares}
}
\description{
glance methods always return either a one-row data frame, or NULL
}

