% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-examples.R
\docType{data}
\name{lobelia}
\alias{lobelia}
\title{Lobelia urens seeds data}
\format{A data frame with 238 rows and 3 variables:
\describe{
  \item{Day}{The day number}
  \item{Temperature}{The temperature}
  \item{Germination}{The number which germinated}
}}
\usage{
lobelia
}
\description{
This data describes the number of germinating lobelia urens seeds at
different temperatures.
}
\examples{
lobelia
}

