% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlstimedist.R
\docType{package}
\name{nlstimedist}
\alias{nlstimedist}
\alias{nlstimedist-package}
\title{Fit the time-course of biological phenomena}
\description{
\pkg{nlstimedist} fits a biologically meaningful distribution function to
time-sequence data (phenology), estimates parameters to draw the cumulative
distribution function and probability density function and calculates
standard statistical moments and percentiles.
}
