% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timedist.R
\name{tdData}
\alias{tdData}
\title{Prepare \code{nlstimedist} data}
\usage{
tdData(data, x, y)
}
\arguments{
\item{data}{The raw data to be cleaned.}

\item{x}{The time variable.}

\item{y}{The number of events.}
}
\description{
The data for \code{nlstimedist} needs to be in a particular format. This
function prepares the data for the model.
}

