% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{tdPDF}
\alias{tdPDF}
\title{Calculate the PDF}
\usage{
tdPDF(x, S = 1, r, c, t)
}
\arguments{
\item{x}{Points at which to calculate the the pdf.}

\item{S}{Unknown}

\item{r, c, t}{Parameter values within the model.}
}
\value{
A vector of values from the pdf.
}
\description{
Calculate values of the probability density function.
}

