% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss.R
\name{tdRSS}
\alias{tdRSS}
\title{Calculate the corrected residual sum of squares}
\usage{
tdRSS(model)
}
\arguments{
\item{model}{An object of class \code{timedist}.}
}
\value{
A single value.
}
\description{
Calculate the corrected residual sum of squares for a model of
  class \code{timedist}.
}

