% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R, R/timedist.R
\docType{package}
\name{timedist}
\alias{timedist}
\alias{timedist-package}
\title{Fit the time-course of biological phenomena}
\usage{
timedist(data, x, y, r, c, t, ...)
}
\arguments{
\item{data}{The data to be included in the model.}

\item{x, y}{The x and y values in the data, where the y values are the
proportions.}

\item{r, c, t}{The starting parameters for the model.}

\item{...}{Additional parameters to be passed to
\code{\link[minpack.lm]{nlsLM}}.}
}
\description{
\pkg{timedist} allows you to...

Fit the Franco model
}
\details{
The \code{\link[minpack.lm]{nlsLM}} function is used instead of the
  \code{\link[stats]{nls}} function in order to use the Levenberg-Marquardt
  algorithm because...
}

