% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-examples.R
\docType{data}
\name{trees}
\alias{trees}
\title{Leafing phenology of tilia cordata}
\format{A data frame with 34 rows and 2 variables:
\describe{
  \item{Day}{The day number}
  \item{Trees}{The number of trees}
}}
\usage{
trees
}
\description{
This data describes the leafing phenology of lime trees (tilia cordata).
}
\examples{
trees
}

