% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-examples.R
\docType{data}
\name{lobelia}
\alias{lobelia}
\title{Lobelia urens seeds data}
\format{
A data frame with 231 rows and 3 variables:
\describe{
\item{Day}{The day number}
\item{Temperature}{The temperature}
\item{Germination}{The number which germinated}
}
}
\usage{
lobelia
}
\description{
This data describes the number of germinating lobelia urens seeds at different temperatures.
}
\details{
The total numbers which failed to germinate are 59, 52, 35, 22, 10, 7 and 12 for temperatures 9.8, 12.5, 16.7, 20.2,
24.3, 28.5 and 32.0, respectively.
}
\examples{
lobelia
}
