% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timedist.R
\name{tdData}
\alias{tdData}
\title{Prepare \code{nlstimedist} data}
\usage{
tdData(data, x, y, group = NULL)
}
\arguments{
\item{data}{A \code{data.frame}. The raw data to be cleaned.}

\item{x}{\code{character(1)}. The time variable.}

\item{y}{\code{character(1)}. The number of events.}

\item{group}{\code{character(1)}. The run numbers. This is \code{NULL} by default if you are only using the function for one
run.}
}
\value{
A \code{data.frame} of the cleaned data to be supplied to the \code{\link[=timedist]{timedist()}} function.
}
\description{
The data for \code{nlstimedist} needs to be in a particular format. This function prepares the data for the model.
}
\examples{
tdData(tilia, x = "Day", y = "Trees")
tdData(lobelia, x = "Day", y = "Germination", group = "Temperature")

}
