% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{tdMoments}
\alias{tdEntropy}
\alias{tdKurtosis}
\alias{tdMean}
\alias{tdMoments}
\alias{tdSkew}
\alias{tdVariance}
\title{Calculate moments for the fitted td model}
\usage{
tdMoments(object, ...)

tdMean(object, upper = Inf, ...)

tdVariance(object, upper = Inf, ...)

tdSkew(object, upper = Inf, ...)

tdKurtosis(object, upper = Inf, alternative = FALSE, ...)

tdEntropy(object, upper = NULL, ...)
}
\arguments{
\item{object}{A model of class \code{td}.}

\item{...}{Additional arguments to be passed to \code{\link[stats]{integrate}}}

\item{upper}{Defaults to \code{Inf} for all moment functions except for entropy which uses
\code{t * 10} by default.}

\item{alternative}{An alternative calculation method}
}
\value{
A single value, or in the case of \code{tdMoments}, a \code{data.frame} of values.
}
\description{
Individual functions are provided as well as a wrapper to calculate the moments for
your fitted model.
}

