% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{tdPDF}
\alias{tdCDF}
\alias{tdPDF}
\title{Calculate the PDF and CDF}
\usage{
tdPDF(x, S = 1, r, c, t)

tdCDF(x, S = 1, r, c, t)
}
\arguments{
\item{x}{Points at which to calculate the the pdf.}

\item{S}{Scaling factor for the PDF.}

\item{r, c, t}{Parameter values within the model.}
}
\value{
A vector of values from the pdf.

A vector of values from the cdf.
}
\description{
Calculate values of the probability density function.

Calculate values of the cumulative distribution function
}

