% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{tdPDF}
\alias{tdPDF}
\alias{tdCDF}
\title{Calculate the PDF and CDF}
\usage{
tdPDF(x, r, c, t, S = 1)

tdCDF(x, r, c, t, S = 1)
}
\arguments{
\item{x}{\code{numeric(n)}. Points at which to calculate the the PDF.}

\item{r, c, t}{\code{numeric(1)}. Parameter values within the model.}

\item{S}{\code{numeric(1)}. Scaling factor for the PDF.}
}
\value{
A vector of values from the PDF or CDF.
}
\description{
Calculate values of the probability density function (PDF) and the cumulative distribution function (CDF).
}
\seealso{
\code{\link[=tdPdfPlot]{tdPdfPlot()}}, \code{\link[=tdCdfPlot]{tdCdfPlot()}}
}
