% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss.R
\name{tdRSS}
\alias{tdRSS}
\title{Calculate the corrected residual sum of squares}
\usage{
tdRSS(model)
}
\arguments{
\item{model}{An object of class \code{timedist}.}
}
\value{
\code{numeric(1)}.
}
\description{
Calculate the corrected residual sum of squares for a \code{timedist} model.
}
\examples{
tdTilia <- tdData(tilia, x = "Day", y = "Trees")
model <- timedist(data = tdTilia, x = "Day", y = "propMax", r = 0.1, c = 0.5, t = 120)
model
tdRSS(model)

}
