% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R, R/timedist.R
\docType{package}
\name{timedist}
\alias{timedist}
\alias{timedist-package}
\title{Fit...}
\usage{
timedist(data, x, y, r, c, t, ...)
}
\arguments{
\item{data}{The data to be included in the model.}

\item{...}{Additional parameters to be passed to \code{\link[minpack.lm]{nlsLM}}.}

\item{start}{A list of starting parameters for \code{r}, \code{c} and \code{t}.}
}
\description{
\pkg{timedist} allows you to

Fit the Franco model
}
\details{
The \code{\link[minpack.lm]{nlsLM}}function is used instead of the
  \code{\link[stats]{nls}} function in order to use the Levenberg-Marquardt algorithm because...
}

